feat: improve ImageCropper
