+++ SwiftUIExample/MantisSwiftUIExample/ImageCropper.swift
@@ -8,18 +8,18 @@
 import Mantis
 import SwiftUI
 
-enum ImageCropperType {
+enum MantisImageCropperType {
     case normal
     case noRotaionDial
     case noAttachedToolbar
 }
 
 struct ImageCropper: UIViewControllerRepresentable {
+    let config: Mantis.Config
+    
     @Binding var image: UIImage?
-    @Binding var cropShapeType: Mantis.CropShapeType
-    @Binding var presetFixedRatioType: Mantis.PresetFixedRatioType
-    @Binding var type: ImageCropperType
     @Binding var transformation: Transformation?
+    @Binding var cropInfo: CropInfo?
     
     @Environment(\.presentationMode) var presentationMode
     
@@ -46,47 +46,13 @@ struct ImageCropper: UIViewControllerRepresentable {
     }
     
     func makeUIViewController(context: Context) -> UIViewController {
-        switch type {
-        case .normal:
-            return makeNormalImageCropper(context: context)
-        case .noRotaionDial:
-            return makeImageCropperHiddingRotationDial(context: context)
-        case .noAttachedToolbar:
-            return makeImageCropperWithoutAttachedToolbar(context: context)
-        }
-    }
-    
-    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
-        
-    }
-}
-
-extension ImageCropper {
-    func makeNormalImageCropper(context: Context) -> UIViewController {
-        var config = Mantis.Config()
-        config.cropViewConfig.cropShapeType = cropShapeType
-        config.presetFixedRatioType = presetFixedRatioType
         let cropViewController = Mantis.cropViewController(image: image!,
                                                            config: config)
         cropViewController.delegate = context.coordinator
         return cropViewController
     }
     
-    func makeImageCropperHiddingRotationDial(context: Context) -> UIViewController {
-        var config = Mantis.Config()
-        config.cropViewConfig.showAttachedRotationControlView = false
-        let cropViewController = Mantis.cropViewController(image: image!, config: config)
-        cropViewController.delegate = context.coordinator
-
-        return cropViewController
-    }
-    
-    func makeImageCropperWithoutAttachedToolbar(context: Context) -> UIViewController {
-        var config = Mantis.Config()
-        config.showAttachedCropToolbar = false
-        let cropViewController: CustomViewController = Mantis.cropViewController(image: image!, config: config)
-        cropViewController.delegate = context.coordinator
-
-        return UINavigationController(rootViewController: cropViewController)
+    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
+        
     }
 }
